+++ BESMIKWebApp/NTier/BESMIK.ViewModel/Spending/SpendingViewModel.cs
@@ -12,7 +12,7 @@ namespace BESMIK.ViewModel.Spending
     public class SpendingViewModel:BaseViewModel
     {
         public SpendingType SpendingType { get; set; }
-        public int Sum { get; set; }
+        public float Sum { get; set; }
         public SpendingCurrency SpendingCurrency { get; set; }
         public SpendingStatus SpendingStatus { get; set; }
         public DateOnly SpendingRequestDate { get; set; }
